- Memory and power consumption profiled and somewhat reduced. - AWSS3TransferManager sucks - threads leak. Will try to replace it after this commit. - Deprecated optimizely use. - Deprecated mixpanel use. - Deprecated iRate. - Using 6 download workers and 2 render workers.

+++ HSDK/update.py
@@ -1,8 +1,26 @@
 import shutil
+import os
+import sys
 
+def print_description():
+	print """update.py <env>
+	Where <env> may be:
+	debug   - Will copy from the debug build folder.
+	release - Will copy from the release build folder.
+	"""
 
 if __name__ == '__main__':
-	print "will update framework files"
+	try:
+		env = sys.argv[1]
+	except:
+		env = None
+
+	if env != "debug" and env != "release":
+		print_description()
+		sys.exit(0)
+
+	print "will update framework files for %s sdk build" % env
+
 	try:
 		shutil.rmtree("HomageSDKCore.framework")
 	except:
@@ -13,10 +31,12 @@ if __name__ == '__main__':
 	except:
 		pass
 
-	# DEBUG
-	shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Debug-iphoneos/HomageSDKCore.framework", "HomageSDKCore.framework")
-	shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Debug-iphoneos/HomageSDKFlow.framework", "HomageSDKFlow.framework")
+	if env == "debug":
+		# DEBUG
+		shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Debug-iphoneos/HomageSDKCore.framework", "HomageSDKCore.framework")
+		shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Debug-iphoneos/HomageSDKFlow.framework", "HomageSDKFlow.framework")
 
-	# RELEASE
-	# shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Release-iphoneos/HomageSDKCore.framework", "HomageSDKCore.framework")
-	# shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Release-iphoneos/HomageSDKFlow.framework", "HomageSDKFlow.framework")
+	else:
+		# RELEASE
+		shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Release-iphoneos/HomageSDKCore.framework", "HomageSDKCore.framework")
+		shutil.copytree("/Users/aviv/Library/Developer/Xcode/DerivedData/HomageSDK-frvsrndvdzevnofmbzqzgumyngkl/Build/Products/Release-iphoneos/HomageSDKFlow.framework", "HomageSDKFlow.framework")
